Code samples - Books
Books is a program that reads a list of books as well as a list of users
and their ratings of the corresponding books. Users are able to search
and sort through the list based on different parameters. It also allows
the user to rate their own books as well as get recommendations based on
other users' ratings.