Code Samples - Chat Room
This program includes a chat server as well as a chat client. This chat
room uses socket communications and multithreading to manage different
clients. Usernames and passwords are saved by the server in a serial
file. Each client is given a thread and is able to communicate both
publicly and privately with other clients. Messages are sent using
objects, allowing picture functionality. Clients are given a GUI that
shows previous messages, as well as users who are logged in and out.