Game
This is the source code for the game that I showed you. Most of the work
was done in game.c, MMA8451.c, and tasks.c. Many of these files are
libraries from the sound and display API. This game is run using a
scheduler to ensure that tasks are assigned priorities and run
accordingly. With the limited resources, speed and simplicity are
important to make the game run smoothly. To do so, I used compiler
optimizations and math approximations to save clock cycles.